refactor(auth): colocate route-specific forms
