+++ src/app/(main)/(auth)/change-password/_components/UpdatePasswordForm.tsx
@@ -1,50 +1,58 @@
-"use client";
+import { Button, TypographyH1, TypographyMuted } from "@/ui/components/atoms";
 
-import { useActionState, useEffect } from "react";
-import { useRouter, useSearchParams } from "next/navigation";
-import { toast } from "sonner";
-
-import { updateUserPassword, clearUserEmailCookie } from "@/actions";
-import { hasFieldErrors } from "@/core/utils";
+import { TextField } from "@/ui/components/organisms";
+import Link from "next/link";
+import { getFieldError } from "@/core/utils";
 import type { APIResponse } from "@/core/domain";
-import { UpdatePasswordForm as PureUpdatePasswordForm } from "@/ui/components/organisms";
 import { routes } from "@/core/domain";
 
-export function UpdatePasswordForm() {
-  const router = useRouter();
-  const token = useSearchParams().get("t") ?? "";
-  const [state, action, pending] = useActionState<
-    APIResponse<{ message: string }>,
-    FormData
-  >(updateUserPassword, null);
+interface UpdatePasswordFormProps {
+  action: (formData: FormData) => void;
+  pending: boolean;
+  state: APIResponse<{ message: string }> | null;
+  token: string;
+}
 
-  useEffect(() => {
-    if (!state) return;
-    if (state.success === true) {
-      toast.message(state.data.message);
-      return router.push(routes.login);
-    } else {
-      if (!hasFieldErrors(state.error)) {
-        toast.error(state.error.message);
-        router.push(routes.login);
-      }
-    }
-  }, [state, router]);
-
-  useEffect(() => {
-    return () => {
-      clearUserEmailCookie().catch((error) => {
-        console.debug("Cookie deletion failed during cleanup:", error);
-      });
-    };
-  }, []);
+export function UpdatePasswordForm({
+  action,
+  pending,
+  state,
+  token,
+}: UpdatePasswordFormProps) {
+  const passwordError = getFieldError(state, "password");
+  const confirmPasswordError = getFieldError(state, "confirmPassword");
 
   return (
-    <PureUpdatePasswordForm
-      action={action}
-      pending={pending}
-      state={state}
-      token={token}
-    />
+    <div className="space-y-6">
+      <div className="space-y-2 text-center lg:text-left">
+        <TypographyH1 className="text-left text-3xl font-bold font-[var(--font-NotoSerif)]">비밀번호 변경</TypographyH1>
+        <TypographyMuted>변경할 비밀번호를 입력해주세요.</TypographyMuted>
+      </div>
+
+      <form action={action} className="space-y-4">
+        <input name="token" defaultValue={token} hidden />
+
+        <TextField id="password" name="password" type="password" placeholder="••••••••" required error={passwordError}>
+          비밀번호
+        </TextField>
+
+        <TextField id="confirmPassword" name="confirmPassword" type="password" placeholder="••••••••" required error={confirmPasswordError}>
+          비밀번호 확인
+        </TextField>
+
+        <Button type="submit" className="w-full" size="lg">
+          비밀번호 변경 {pending ? "중" : "완료"}
+        </Button>
+      </form>
+
+      <div className="space-y-2 text-center">
+        <Link
+          href={routes.login}
+          className="text-muted-foreground hover:text-foreground inline-block text-sm transition-colors"
+        >
+          로그인으로 돌아가기
+        </Link>
+      </div>
+    </div>
   );
 }
