--- src/app/(main)/(auth)/change-password/_components/index.tsx
@@ -1,1 +1,0 @@
-export { UpdatePasswordForm } from "./UpdatePasswordForm";
