+++ src/app/(main)/(auth)/change-password/_containers/index.tsx
@@ -1,0 +1,1 @@
+export { UpdatePasswordForm } from "./UpdatePasswordForm";
