--- src/app/(main)/(auth)/find-id/_components/index.tsx
@@ -1,1 +1,0 @@
-export { FindIdForm } from "./FindIdForm";
