+++ src/app/(main)/(auth)/find-id/_containers/index.tsx
@@ -1,0 +1,1 @@
+export { FindIdForm } from "./FindIdForm";
