--- src/app/(main)/(auth)/find-password/_components/index.tsx
@@ -1,1 +1,0 @@
-export { ForgotPasswordForm } from "./ForgotPasswordForm";
