+++ src/app/(main)/(auth)/find-password/_containers/index.tsx
@@ -1,0 +1,1 @@
+export { ForgotPasswordForm } from "./ForgotPasswordForm";
