--- src/app/(main)/(auth)/login/_components/index.tsx
@@ -1,1 +1,0 @@
-export { LoginForm } from "./LoginForm";
