--- src/app/(main)/(auth)/signup/_components/index.tsx
@@ -1,1 +1,0 @@
-export { SignupForm } from "./SignupForm";
