+++ src/app/(main)/(auth)/signup/_containers/index.tsx
@@ -1,0 +1,1 @@
+export { SignupForm } from "./SignupForm";
