+++ src/app/(main)/(auth)/signup/page.tsx
@@ -1,4 +1,4 @@
-import { SignupForm } from "./_components";
+import { SignupForm } from "./_containers";
 
 export default function SignupPage() {
   return <SignupForm />;
