+++ src/ui/components/organisms/index.ts
@@ -12,14 +12,11 @@ export * from "./CreateGuestbookForm";
 export * from "./DeleteGuestbookForm";
 export * from "./EcommerceHero";
 export * from "./ErrorFallback";
-export * from "./FindIdForm";
-export * from "./ForgotPasswordForm";
 export * from "./FormField";
 export * from "./ImageField";
 export * from "./ImagePreviewItem";
 export * from "./ImagesSection";
 export * from "./LiveDemoSection";
-export * from "./LoginForm";
 export * from "./MobileNav";
 export * from "./ParentsInfoSection";
 export * from "./PersonValueCard";
@@ -38,7 +35,5 @@ export * from "./RatingStars";
 export * from "./SelectField";
 export * from "./SidebarNavItem";
 export * from "./SidebarToggle";
-export * from "./SignupForm";
 export * from "./SwitchField";
 export * from "./TextField";
-export * from "./UpdatePasswordForm";
